Restore ability to create user mappings (#80)
In bff2f71e4deff4e522506aa9af8d28a8422faed4 cmd_nmap was refactored to
generically handle either user_mappings or user_mappings_noremap by
adding a new parameter to the function, key.  However, this was shadowed
later in the function when the mapping was being created or modified.

Fix this with a parameter name change and make a regression less likely
by using an enum to represent the different mapping types.